+++ src/components/Icon/Icon.js
@@ -1,5 +1,4 @@
 // @flow
-// TODO: move to uikit
 import * as React from 'react'
 import { css, cx } from 'emotion';
 // import { iconSize } from '../../variables';