+++ src/components/Icon/icons/IconCreateFile/index.js
@@ -1,0 +1,18 @@
+// @flow
+import * as React from 'react';
+import { css, cx } from 'emotion';
+import { Icon, type GenericIconProps } from '../../Icon';
+import image from './create-file.svg';
+
+const styles = css`
+  width: 12px;
+  height: 12px;
+`;
+
+export const IconCreateFile = ({ className, onClick }: GenericIconProps) => (
+  <Icon
+    className={cx(styles, className)}
+    glyph={image}
+    onClick={onClick}
+  />
+);