+++ CHANGELOG.md
@@ -4,13 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.6.0] - 2019-12-05
 
 - Updates in Modal: new `fit` prop, new behavior with content overload and scrollbars
 
 - Added Tooltip component
 
 - Export text styles from Text component
+
+- FormField and LabeledInput moved from cartridge
 
 ## [0.5.0] - 2019-11-27